Split up ed25519 into multiple files.
