+++ e2e/Libs/Ristretto/ed25519.py
@@ -9,94 +9,8 @@ import hashlib
 import gmpy2
 from gmpy2 import mpz
 
-import e2e.Libs.Ristretto.Ed25519Reference as ed
-
-class Ed25519Point:
-  underlying: List[mpz]
-
-  def __init__(
-    self,
-    point: Union[List[mpz], bytes]
-  ) -> None:
-    if isinstance(point, bytes):
-      self.underlying = ed.decodepoint(point)
-    else:
-      self.underlying = point
-
-  def __add__(
-    self,
-    other: 'Ed25519Point'
-  ) -> 'Ed25519Point':
-    x1: mpz = self.underlying[0]
-    y1: mpz = self.underlying[1]
-    x2: mpz = other.underlying[0]
-    y2: mpz = other.underlying[1]
-    x3: mpz = ((x1 * y2) + (x2 * y1)) * ed.inv(ed.ONE + (ed.d * x1 * x2 * y1 * y2))
-    y3: mpz = ((y1 * y2) + (x1 * x2)) * ed.inv(ed.ONE - (ed.d * x1 * x2 * y1 * y2))
-    return Ed25519Point([x3 % ed.q, y3 % ed.q])
-
-  def __mul__(
-    self,
-    scalar: 'Ed25519Scalar'
-  ) -> 'Ed25519Point':
-    if scalar.underlying == ed.ZERO:
-      return Ed25519Point([ed.ZERO, ed.ONE])
-    res: Ed25519Point = self * Ed25519Scalar(scalar.underlying // ed.TWO)
-    res = res + res
-    if scalar.underlying & ed.ONE:
-      #pylint: disable=arguments-out-of-order
-      res = res + self
-    return res
-
-  def serialize(
-    self
-  ) -> bytes:
-    res: bytearray = bytearray(gmpy2.to_binary(self.underlying[1])[2:].ljust(32, b"\0"))
-    res[-1] = (((res[-1] << 1) & 255) >> 1) | ((int(self.underlying[0]) & 1) << 7)
-    return bytes(res)
-
-BASEPOINT: Ed25519Point = Ed25519Point(ed.B)
-
-class Ed25519Scalar:
-  underlying: mpz
-
-  def __init__(
-    self,
-    scalar: Union[mpz, int, bytes, bytearray]
-  ) -> None:
-    if isinstance(scalar, bytearray):
-      scalar = bytes(scalar)
-    if isinstance(scalar, bytes):
-      self.underlying = gmpy2.from_binary(b"\1\1" + scalar)
-    elif isinstance(scalar, int):
-      self.underlying = mpz(scalar)
-    else:
-      self.underlying = scalar
-    self.underlying = self.underlying % ed.l
-
-  def serialize(
-    self
-  ) -> bytes:
-    return gmpy2.to_binary(self.underlying)[2:].ljust(32, b"\0")
-
-  def __add__(
-    self,
-    other: 'Ed25519Scalar'
-  ) -> 'Ed25519Scalar':
-    return Ed25519Scalar(self.underlying + other.underlying)
-
-  def __mul__(
-    self,
-    scalar: 'Ed25519Scalar'
-  ) -> 'Ed25519Scalar':
-    return Ed25519Scalar(self.underlying * scalar.underlying)
-
-  def toPoint(
-    self
-  ) -> Ed25519Point:
-    return BASEPOINT * self
-
-MODULUS: Ed25519Scalar = Ed25519Scalar(ed.l)
+from e2e.Libs.Ristretto.Point import Point as Ed25519Point, BASEPOINT
+from e2e.Libs.Ristretto.Scalar import Scalar as Ed25519Scalar, MODULUS
 
 def Hint(
   m: bytes
@@ -121,9 +35,9 @@ def aggregate(
     L = L + key.serialize()
   L = hashlib.blake2b(L).digest()
 
-  res: Ed25519Point = keys[0] * Bint(b"agg" + L + keys[0].serialize())
+  res: Ed25519Point = keys[0] * Bint(b"agg" + L + keys[0].serialize()).underlying
   for k in range(1, len(keys)):
-    res = res + (keys[k] * Bint(b"agg" + L + keys[k].serialize()))
+    res = res + (keys[k] * Bint(b"agg" + L + keys[k].serialize()).underlying)
   return res
 
 def seedToExtendedKey(
@@ -147,14 +61,14 @@ class SigningKey:
       esk = seedToExtendedKey(seed)
     self.scalar: Ed25519Scalar = Ed25519Scalar(esk[:32])
     self.nonce: bytes = esk[32:]
-    self.publicKey: bytes = (BASEPOINT * self.scalar).serialize()
+    self.publicKey: bytes = self.scalar.toPoint().serialize()
 
   def sign(
     self,
     msg: bytes
   ) -> bytes:
     r: Ed25519Scalar = Hint(self.nonce + msg)
-    R: bytes = (BASEPOINT * r).serialize()
+    R: bytes = r.toPoint().serialize()
     k: Ed25519Scalar = Hint(R + self.publicKey + msg)
     return R + (r + (k * self.scalar)).serialize()
